+++ solarboat/src/main/java/hu/schdesign/solarboat/model/Boat/Motor.java
@@ -11,11 +11,11 @@ public class Motor implements CsvPrintable {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     private Long id;
-    private final int RpM;
-    private final int temp;
+    private final double RpM;
+    private final double temp;
     public static final char CSV_SEPARATOR = ';';
 
-    public Motor(@JsonProperty("rpM") int rpM, @JsonProperty ("temp") int temp) {
+    public Motor(@JsonProperty("rpM") double rpM, @JsonProperty ("temp") double temp) {
         RpM = rpM;
         this.temp = temp;
     }
@@ -25,11 +25,11 @@ public class Motor implements CsvPrintable {
         this.temp = 0;
     }
 
-    public int getRpM() {
+    public double getRpM() {
         return RpM;
     }
 
-    public int getTemp() {
+    public double getTemp() {
         return temp;
     }
